pyproject.toml
[project]
name = "behavioralsignals"
version = "0.1.0"
description = "Python SDK for Behavioral Signals API"
readme = "README.md"
authors = [
    { name = "Nikos Antoniou", email = "nikos.antoniou@behavioralsignals.com" }
]
requires-python = ">=3.10"
dependencies = [
    "dotenv>=0.9.9",
    "grpcio>=1.64.0",
    "grpcio-tools>=1.64.0",
    "pydub>=0.25.1",
    "requests>=2.32.4",
]


[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
